+++ src/main/java/ru/obukhov/trader/trading/model/BackTestPosition.java
@@ -2,5 +2,5 @@ package ru.obukhov.trader.trading.model;
 
 import java.math.BigDecimal;
 
-public record BackTestPosition(String ticker, BigDecimal price, int quantity) {
+public record BackTestPosition(String ticker, BigDecimal price, long quantity) {
 }
